updated home readme
We have never perfected the selling of changesets - even when we think it's awesome, we struggle to tell people why. (the fact that I always tell people what it does, not what problems it solves wasn't helping)

I'm trying to fix that, and make a reference to help me fix it when talking to other people in the future.

This work is inspired by me being fed-up with this problem.