fix: handle errors when checking divergent branches
I just ran into the issue where my CI environment was performing a
shallow clone, and so it didn't have all the refs in its local copy.

When I used the `getReleasePlan(cwd, MY_CUSTOM_REF)` function, it
silently failed and returned an empty release plan.
It took me a while to figure out what the problem was!

This change adds error handling when we call out to `git`, so if the ref
itself doesn't actually exist then the function throws.