Helpful note about `status` command
I was trying to use the `status` command to generate a JSON output file for using in CI to report back on which version items were being published. I kept trying to get the status after the `version` and `publish` commands as that seemed like the logical state I would be wanting to get status. Hopefully this note helps someone else.