Tweak an example in the docs (#741)
When I first read this doc file, I interpreted the sample changeset file as saying that `@changesets/cli` is a special metadata tag (ie, that the use of "changesets" here was semantically meaningful and not just a dogfooding example). Maybe this was just my confusion but if you agree that this can be confusing, I think this would help.