Fix: use option for only workspace:^ pkgs (#568)
The bumpVersionsWithWorkspaceProtocolOnly was added in commit
12f9a433a6c3ac38f9405fcd77c9108c423d7101 in January 2021 but it never
made its way to the `config` package.

The state of the world before this fix was that `changeset add` would
respect the option and not bump packages that didn't have the
`workspace` range when the option was true, but `changeset version`
would log a warning:

	Package "foo" must depend on the current version of "bar:2.0.0"
	vs "bar:1.0.0"

This fix just stops printing that message.