Update adding-a-changeset.md
The integration with changesets-gitlab points to this documentation resource. 

The link is embedded a comment added by a bot to a gitlab merge-request thread. 

For our monorepo this integration therefore guides people to run npx or yarn but doesn't guide them to use the package manager `pnpm` that is in use within our monorepo.